edit: add-at-end-of-addr is after addr, not inside addr
There was a bug where adding at exactly the end position of an addr would extend the address. It should simply add after it.

I also made the update function significantly more explicit about the cases. This probably adds 1 or 2 redundant comparisons, but it won't matter and makes the logic much easier to reason about.