text: don't store page size; always compute it
If the size of the box changes, the page size changes. Previously we never recomputed it, so the page size was fixed to the box creation size. Now we simply recompute it when needed. We could instead just recompute it on resize, but Box already has too many fields, so I prefer to reduce the number of fields.