text: fix and simplify drawing to images with non-zero Min point
There was a bug with computing tab stops in the case of images with non-zero Min points. But overall, the handling of such images was needlessly complex. Instead of accounting for Min all over the drawing code, we now just assume it's 0,0 until just before a call to draw.Draw, where we then offset by the min point.