ui: some color changes
I found that at full screen, the low contrast between the padding color and the background color of sheets and tags was hurting my eyes. Better to have a single pixel of black line separating things.

I also changed the body background to be less bright. (I used the same color as my custom Acme color scheme, which I've been using for years, and it doesn't make my eyes tired event after long stretches of coding).

I also change the column background color to white. This was to draw a bit more contrast in the colum and tag handles, which use inverses of eachothers' colors.