re1, edit: add re1.Union
It returns a Regexp matches the union of the given regular expressions. This is similar to | of the regexps, but without re-numbering the capture groups. The intention is for creating lexers/scanners by composing separate Regexps, one for each token type.

To distinguish which component matched, the compile Opts now has an ID field. This is an integer, specified by the called. The last element of the slice returned by Find on match is the ID of the matching component Regexp.